requirements.txt
aiohttp>=3.8.5
discord.py>=2.0.1
discord.py[voice]
gtts>=2.5.0
langdetect>=1.0.9
python-dotenv>=0.21.0
pyttsx3>=2.90
PyYAML>=6.0.1
requests>=2.31.0
openai>=1.11.1
duckduckgo-search>=4.1.1
groq>=0.4.1

# Agent dependencies
langchain>=0.3.25
langchain-groq>=0.3.2
langchain-community>=0.3.24
langchain-core>=0.3.59
langchain-anthropic>=0.1.1
langchain-openai>=0.1.1

# RAG dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# MCP dependencies
langchain-mcp>=0.1.3
langchain-mcp-adapters>=0.0.3
mcp>=0.0.12

# Token counting
tiktoken>=0.6.0

# Monitoring dependencies
matplotlib>=3.7.1
aiofiles>=23.2.1
numpy>=1.24.3

# Image processing
pillow>=10.0.0

# Utility
python-dateutil>=2.8.2
pydantic>=2.0.0